feat: move api keys on the dashboard to new tab
