+++ dashboard/src/components/OrgTabs.tsx
@@ -35,6 +35,13 @@ export const OrgTabs = () => {
       >
         Settings
       </A>
+      <A
+        href={`/dashboard/${userContext.selectedOrganizationId?.()}/api-keys`}
+        activeClass="border-b-2 -mb-[1px] border-magenta-500"
+        class="hover:text-fuchsia-800"
+      >
+        API Keys
+      </A>
     </div>
   );
 };
