+++ dashboard/src/pages/Dashboard/Dataset/DatasetStart.tsx
@@ -1,38 +1,14 @@
-import {
-  For,
-  Show,
-  createEffect,
-  createMemo,
-  createSignal,
-  on,
-  onCleanup,
-  useContext,
-} from "solid-js";
-import {
-  ApiKeyDTO,
-  fromI32ToApiKeyRole,
-  fromI32ToUserRole,
-} from "../../../types/apiTypes";
-import { UserContext } from "../../../contexts/UserContext";
 import { BiRegularInfoCircle, BiRegularLinkExternal } from "solid-icons/bi";
 import CreateChunkRequest from "../../../components/CreateChunkRequest.md";
 import HybridSearchReqeust from "../../../components/HybridSearchRequest.md";
-import { DatasetContext } from "../../../contexts/DatasetContext";
-import { ApiKeyGenerateModal } from "../../../components/ApiKeyGenerateModal";
-import { FaRegularClipboard, FaRegularTrashCan } from "solid-icons/fa";
-import { formatDate } from "../../../formatters";
 import { BuildingSomething } from "../../../components/BuildingSomething";
 import { BsMagic } from "solid-icons/bs";
 import { AddSampleDataModal } from "../../../components/DatasetExampleModal";
 
 export const DatasetStart = () => {
-  const api_host = import.meta.env.VITE_API_HOST as unknown as string;
-
   const userContext = useContext(UserContext);
   const datasetContext = useContext(DatasetContext);
 
-  const [apiKeys, setApiKeys] = createSignal<ApiKeyDTO[]>([]);
-  const [openModal, setOpenModal] = createSignal<boolean>(false);
   const [openSampleDataModal, setOpenSampleDataModal] =
     createSignal<boolean>(false);
 
@@ -46,69 +22,6 @@ export const DatasetStart = () => {
     const dataset = datasetContext.dataset?.();
     if (!dataset) return null;
     return dataset;
-  });
-
-  const currentUserRole = createMemo(() => {
-    const selectedOrgId = userContext.selectedOrganizationId?.();
-    if (!selectedOrgId) return 0;
-    return (
-      userContext
-        .user?.()
-        ?.user_orgs.find(
-          (user_org) => user_org.organization_id === selectedOrgId,
-        )?.role ?? 0
-    );
-  });
-
-  const getApiKeys = (abortController: AbortController) => {
-    void fetch(`${api_host}/user/api_key`, {
-      method: "GET",
-      credentials: "include",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      signal: abortController.signal,
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        setApiKeys(data);
-      });
-  };
-
-  const deleteApiKey = (id: string) => {
-    void fetch(`${api_host}/user/api_key/${id}`, {
-      method: "DELETE",
-      credentials: "include",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    }).then((resp) => {
-      if (resp.ok) {
-        getApiKeys(new AbortController());
-      }
-    });
-  };
-
-  createEffect(
-    on(openModal, () => {
-      const abortController = new AbortController();
-
-      getApiKeys(abortController);
-
-      onCleanup(() => {
-        abortController.abort();
-      });
-    }),
-  );
-
-  createEffect(() => {
-    const abortController = new AbortController();
-
-    getApiKeys(abortController);
-
-    onCleanup(() => {
-      abortController.abort();
-    });
   });
 
   return (
@@ -140,131 +53,6 @@ export const DatasetStart = () => {
               </button>
             </div>
             <BuildingSomething />
-            <div class="flex flex-col gap-2">
-              <div class="flex items-center space-x-3">
-                <p class="block text-sm font-medium">
-                  {curDataset()?.name} dataset id:{" "}
-                </p>
-                <p class="w-fit text-sm">{curDataset()?.id}</p>
-                <button
-                  class="text-sm underline"
-                  onClick={() => {
-                    void navigator.clipboard.writeText(curDataset()?.id ?? "");
-                    window.dispatchEvent(
-                      new CustomEvent("show-toast", {
-                        detail: {
-                          type: "info",
-                          title: "Copied",
-                          message: "Dataset ID copied to clipboard",
-                        },
-                      }),
-                    );
-                  }}
-                >
-                  <FaRegularClipboard />
-                </button>
-              </div>
-              <div class="flex items-center space-x-3">
-                <p class="block text-sm font-medium">
-                  {selectedOrganization()?.name} org id:
-                </p>
-                <p class="w-fit text-sm">{selectedOrganization()?.id}</p>
-                <button
-                  class="text-sm underline"
-                  onClick={() => {
-                    void navigator.clipboard.writeText(
-                      selectedOrganization()?.id ?? "",
-                    );
-                    window.dispatchEvent(
-                      new CustomEvent("show-toast", {
-                        detail: {
-                          type: "info",
-                          title: "Copied",
-                          message: "Organization ID copied to clipboard",
-                        },
-                      }),
-                    );
-                  }}
-                >
-                  <FaRegularClipboard />
-                </button>
-              </div>
-              <div class="mt-4 border-t pt-4">
-                <div class="block font-medium">API Keys</div>
-                <Show when={apiKeys().length > 0}>
-                  <div class="mb-1">
-                    <div class="inline-block min-w-full overflow-hidden rounded-md border-[0.5px] border-neutral-300 py-2 align-middle">
-                      <table class="min-w-full divide-y divide-gray-300">
-                        <thead>
-                          <tr>
-                            <th
-                              scope="col"
-                              class="py-3.5 pl-4 pr-3 text-left text-sm font-semibold text-gray-900 sm:pl-6 lg:pl-8"
-                            >
-                              Name
-                            </th>
-                            <th
-                              scope="col"
-                              class="px-3 py-3.5 text-left text-sm font-semibold text-gray-900"
-                            >
-                              Created At
-                            </th>
-                            <th class="px-3 py-3.5 text-left text-sm font-semibold text-gray-900">
-                              Perms
-                            </th>
-                          </tr>
-                        </thead>
-                        <tbody class="divide-y divide-gray-200 bg-white">
-                          <For each={apiKeys()}>
-                            {(apiKey) => (
-                              <tr>
-                                <td class="whitespace-nowrap py-4 pl-4 pr-3 text-sm text-gray-900 sm:pl-6 lg:pl-8">
-                                  {apiKey.name}
-                                </td>
-                                <td class="px-3 py-3.5 text-left text-sm text-gray-900">
-                                  {formatDate(new Date(apiKey.created_at))}
-                                </td>
-                                <td class="px-3 py-3.5 text-left text-sm text-gray-900">
-                                  {apiKey.role > 0
-                                    ? fromI32ToUserRole(currentUserRole())
-                                    : fromI32ToApiKeyRole(
-                                        apiKey.role,
-                                      ).toString()}
-                                </td>
-                                <td class="px-3 py-3.5 text-left text-sm text-gray-900">
-                                  <button>
-                                    <FaRegularTrashCan
-                                      onClick={(e) => {
-                                        e.preventDefault();
-                                        deleteApiKey(apiKey.id);
-                                      }}
-                                    />
-                                  </button>
-                                </td>
-                              </tr>
-                            )}
-                          </For>
-                        </tbody>
-                      </table>
-                    </div>
-                  </div>
-                </Show>
-
-                <button
-                  type="button"
-                  classList={{
-                    "inline-flex mt-2 justify-center rounded-md bg-magenta-500 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-magenta-700 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-magenta-900":
-                      true,
-                  }}
-                  onClick={(e) => {
-                    e.preventDefault();
-                    setOpenModal(true);
-                  }}
-                >
-                  Create New API Key +
-                </button>
-              </div>
-            </div>
           </section>
           <section
             class="flex-col gap-4 border bg-white px-4 py-6 shadow sm:overflow-hidden sm:rounded-md sm:p-6 lg:col-span-2"
@@ -341,10 +129,6 @@ export const DatasetStart = () => {
           </section>
         </div>
       </main>
-      <ApiKeyGenerateModal
-        openModal={openModal}
-        closeModal={() => setOpenModal(false)}
-      />
       <AddSampleDataModal
         openModal={openSampleDataModal}
         closeModal={() => setOpenSampleDataModal(false)}
